# perf(serde_v8): optimize ByteString deserialization