--- a/serde_v8/src/error.rs
+++ b/serde_v8/src/error.rs
@@ -19,6 +19,7 @@ pub enum Error {
   ExpectedBuffer,
 
   ExpectedUtf8,
+  ExpectedLatin1,
 
   LengthMismatch,
 }