Before:
```
test de_bstr_v8_1024_b    ... bench:          67 ns/iter (+/- 1)
test de_bstr_v8_12_b      ... bench:          38 ns/iter (+/- 1)
```
After:
```
test de_bstr_v8_1024_b    ... bench:          50 ns/iter (+/- 2)
test de_bstr_v8_12_b      ... bench:          33 ns/iter (+/- 0)
```